Projects/02_todo_api_project/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
